+++ src/index.ts
@@ -23,8 +23,8 @@ export * from "./widgets/text-editor-field/text-editor-field";
 export * from "./widgets/password-field/password-field";
 export * from "./widgets/dropdown-field/dropdown-field";
 export * from "./widgets/checkbox-field/checkbox-field";
+export * from "./widgets/select-field/select-field";
 
-export * from "../src-old/form/select-field";
 export * from "../src-old/form/file-picker";
 export * from "../src-old/form/file-field";
 export * from "../src-old/form/date-picker";